Unmount a bit more intelligently.
findmnt(8) will let us find the loop device without having to remember
it. umount --recursive will unmount everything without having to track
it all; this is important as some of the package installs mount new
virtual filesystems (naughty naughty).

This is all very Linux-specific, but I think that's unavoidable without
a massive portability layer.

Include a label on the root fs; the boot code needs it.